requirements.txt
streamlit
spacy==3.8.11
pandas
langchain
python-docx
PyPDF2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
fastapi
uvicorn
python-multipart
langchain-google-genai
python-dotenv
pymongo